feat: call different action set based on user scope
