+++ packages/client/src/v2-events/features/events/actions/declare/Review.tsx
@@ -23,14 +23,7 @@ import {
   Text,
   TextInput
 } from '@opencrvs/components'
-import {
-  ActionFormData,
-  ActionType,
-  findActiveActionForm,
-  FormConfig,
-  Scope,
-  SCOPES
-} from '@opencrvs/commons/client'
+import { ActionType, findActiveActionForm } from '@opencrvs/commons/client'
 import { useEventConfiguration } from '@client/v2-events/features/events/useEventConfiguration'
 import { useEventFormData } from '@client/v2-events/features/events/useEventFormData'
 import { useEventMetadata } from '@client/v2-events/features/events/useEventMeta'
@@ -43,8 +36,8 @@ import { FormLayout } from '@client/v2-events/layouts'
 
 // eslint-disable-next-line no-restricted-imports
 import { getScope } from '@client/profile/profileSelectors'
-import { validationErrorsInActionFormExist } from '@client/v2-events/components/forms/validation'
 import { withSuspense } from '@client/v2-events/components/withSuspense'
+import { useReviewActionConfig } from './useReviewActionConfig'
 
 const rejectModalMessages = defineMessages({
   rejectModalCancel: {
@@ -80,252 +73,6 @@ const rejectModalMessages = defineMessages({
   }
 })
 
-const confirmModalMessages = {
-  complete: {
-    declare: {
-      title: {
-        id: 'v2.review.declare.confirmModal.title',
-        defaultMessage: 'Send for review?',
-        description: 'The title for review action modal when declaring'
-      },
-      description: {
-        id: 'v2.review.declare.confirmModal.description',
-        defaultMessage: 'This declaration will be sent for review',
-        description: 'The description for review action modal when declaring'
-      },
-      onConfirm: {
-        id: 'v2.review.declare.confirmModal.confirm',
-        defaultMessage: 'Confirm',
-        description: 'The label for modal confirm button when declaring'
-      },
-      onCancel: {
-        id: 'v2.review.declare.confirmModal.cancel',
-        defaultMessage: 'Cancel',
-        description: 'The label for modal cancel button when declaring'
-      }
-    },
-    validate: {
-      title: {
-        id: 'v2.review.validate.confirmModal.title',
-        defaultMessage: 'Send for approval?',
-        description: 'The title for review action modal when validating'
-      },
-      description: {
-        id: 'v2.review.validate.confirmModal.description',
-        defaultMessage:
-          'This declaration will be sent for approval prior to registration.',
-        description: 'The description for review action modal when validating'
-      },
-      onConfirm: {
-        id: 'v2.review.validate.confirmModal.confirm',
-        defaultMessage: 'Confirm',
-        description: 'The label for modal confirm button when validating'
-      },
-      onCancel: {
-        id: 'v2.review.validate.confirmModal.cancel',
-        defaultMessage: 'Cancel',
-        description: 'The label for modal cancel button when validating'
-      }
-    },
-    register: {
-      title: {
-        id: 'v2.review.register.confirmModal.title',
-        defaultMessage: 'Register?',
-        description: 'The title for review action modal when registering'
-      },
-      description: {
-        id: 'v2.review.register.confirmModal.description',
-        defaultMessage: '‎', // intentionally empty, as the description is not used in v1
-        description: 'The description for review action modal when registering'
-      },
-      onConfirm: {
-        id: 'v2.review.register.confirmModal.confirm',
-        defaultMessage: 'Register',
-        description: 'The label for modal confirm button when registering'
-      },
-      onCancel: {
-        id: 'v2.review.register.confirmModal.cancel',
-        defaultMessage: 'Cancel',
-        description: 'The label for modal cancel button when registering'
-      }
-    }
-  }
-}
-
-const registerMessages = {
-  title: {
-    id: 'v2.review.register.title',
-    defaultMessage: 'Register event',
-    description: 'The title shown when reviewing a record to register'
-  },
-  onConfirm: {
-    id: 'v2.review.register.confirm',
-    defaultMessage: 'Register',
-    description: 'The label for register button of review action'
-  }
-}
-
-const validateMessages = {
-  title: {
-    id: 'v2.review.validate.title',
-    defaultMessage: 'Send for approval',
-    description: 'The title shown when reviewing a record to validate'
-  },
-  onConfirm: {
-    id: 'v2.review.validate.confirm',
-    defaultMessage: 'Send for approval',
-    description: 'The label for review action button when validating'
-  }
-}
-
-const declareMessages = {
-  onConfirm: {
-    id: 'v2.review.declare.confirm',
-    defaultMessage: 'Send for review',
-    description: 'The label for review action button when declaring'
-  }
-}
-
-const reviewMessages = {
-  complete: {
-    register: {
-      title: registerMessages.title,
-      description: {
-        id: 'v2.review.register.description.complete',
-        defaultMessage:
-          'By clicking register, you confirm that the information entered is correct and the event can be registered.',
-        description:
-          'The description for registration action when form is complete'
-      },
-      onConfirm: registerMessages.onConfirm,
-      modal: confirmModalMessages.complete.register
-    },
-    validate: {
-      title: validateMessages.title,
-      description: {
-        id: 'v2.review.validate.description.complete',
-        defaultMessage:
-          'The informant will receive an email with a registration number that they can use to collect the certificate',
-        description: 'The description for validate action when form is complete'
-      },
-      onConfirm: validateMessages.onConfirm,
-      modal: confirmModalMessages.complete.validate
-    },
-    declare: {
-      title: {
-        id: 'v2.review.declare.title.complete',
-        defaultMessage: 'Declaration complete',
-        description:
-          'The title shown when reviewing an incomplete record to declare'
-      },
-      description: {
-        id: 'v2.review.declare.description.complete',
-        defaultMessage:
-          'The informant will receive an email with a registration number that they can use to collect the certificate',
-        description: 'The description for declare action when form is complete'
-      },
-      onConfirm: declareMessages.onConfirm,
-      modal: confirmModalMessages.complete.declare
-    }
-  },
-  incomplete: {
-    register: {
-      title: registerMessages.title,
-      description: {
-        id: 'v2.reviewAction.register.description.incomplete',
-        defaultMessage: 'Please add mandatory information before registering',
-        description:
-          'The description for registration action when form is incomplete'
-      },
-      onConfirm: registerMessages.onConfirm,
-      modal: {}
-    },
-    validate: {
-      title: validateMessages.title,
-      description: {
-        id: 'v2.review.validate.description.incomplete',
-        defaultMessage:
-          'Please add mandatory information before sending for approval',
-        description: 'The description for validate action when form is complete'
-      },
-      onConfirm: validateMessages.onConfirm,
-      modal: {}
-    },
-    declare: {
-      title: {
-        id: 'v2.review.declare.title.incomplete',
-        defaultMessage: 'Declaration incomplete',
-        description:
-          'The title shown when reviewing an incomplete record to declare'
-      },
-      description: {
-        id: 'v2.review.declare.description.complete',
-        defaultMessage:
-          'The informant will receive an email with a tracking ID that they can use to provide the additional mandatory information required for registration',
-        description: 'The description for declare action when form is complete'
-      },
-      onConfirm: declareMessages.onConfirm,
-      modal: {}
-    }
-  }
-}
-function getReviewActionConfig({
-  formConfig,
-  form,
-  metadata,
-  scopes
-}: {
-  formConfig: FormConfig
-  form: ActionFormData
-  metadata?: ActionFormData
-  scopes?: Scope[]
-}) {
-  const incomplete = validationErrorsInActionFormExist(
-    formConfig,
-    form,
-    metadata
-  )
-
-  const isDisabled = incomplete
-    ? !scopes?.includes(SCOPES.RECORD_SUBMIT_INCOMPLETE)
-    : false
-
-  if (scopes?.includes(SCOPES.RECORD_REGISTER)) {
-    return {
-      buttonType: 'positive' as const,
-      incomplete,
-      isDisabled,
-      messages: incomplete
-        ? reviewMessages.incomplete.register
-        : reviewMessages.complete.register
-    }
-  }
-
-  if (scopes?.includes(SCOPES.RECORD_SUBMIT_FOR_APPROVAL)) {
-    return {
-      buttonType: 'positive' as const,
-      incomplete,
-      isDisabled,
-      messages: incomplete
-        ? reviewMessages.incomplete.validate
-        : reviewMessages.complete.validate
-    }
-  }
-
-  if (scopes?.includes(SCOPES.RECORD_DECLARE)) {
-    return {
-      buttonType: 'primary' as const,
-      incomplete,
-      isDisabled,
-      messages: incomplete
-        ? reviewMessages.incomplete.declare
-        : reviewMessages.complete.declare
-    }
-  }
-
-  throw new Error('No valid scope found for the action')
-}
-
 // eslint-disable-next-line no-shadow
 enum REJECT_ACTIONS {
   ARCHIVE,
@@ -345,7 +92,6 @@ export function Review() {
   const [modal, openModal] = useModal()
   const intl = useIntl()
   const { goToHome } = useEventFormNavigation()
-  const declareMutation = events.actions.declare
 
   const [event] = events.getEvent.useSuspenseQuery(eventId)
 
@@ -363,7 +109,7 @@ export function Review() {
 
   const scopes = useSelector(getScope) ?? undefined
 
-  const reviewActionConfiguration = getReviewActionConfig({
+  const reviewActionConfiguration = useReviewActionConfig({
     formConfig,
     form,
     metadata,
@@ -409,12 +155,8 @@ export function Review() {
       />
     ))
     if (confirmedDeclaration) {
-      declareMutation.mutate({
-        eventId: event.id,
-        data: form,
-        transactionId: uuid(),
-        metadata
-      })
+      reviewActionConfiguration.onConfirm(eventId)
+
       goToHome()
     }
   }
