+++ packages/client/src/v2-events/features/events/useEvents/api.ts
@@ -23,10 +23,10 @@ import {
 import { EventDocument, EventIndex } from '@opencrvs/commons/client'
 import { AppRouter, queryClient, utils } from '@client/v2-events/trpc'
 
-export function getLocalEventData(eventId: string) {
-  return queryClient.getQueryData(
-    utils.event.get.queryKey(eventId)
-  ) as EventDocument
+export function findLocalEventData(eventId: string) {
+  return queryClient.getQueryData(utils.event.get.queryKey(eventId)) as
+    | EventDocument
+    | undefined
 }
 
 export function setEventData(id: string, data: EventDocument) {
